+++ RELEASES.md
@@ -7,6 +7,8 @@ Version 0.8.16 (2022-XX-XX)
 * Hyundai Ioniq 5 2022 support thanks to sunnyhaibin!
 * Hyundai Kona Electric 2022 support thanks to sunnyhaibin!
 * Subaru Outback 2020-22 support
+* Honda Odyssey 2021-22 support thanks to csouers!
+* Acura RDX 2022 support thanks to csouers!
 
 Version 0.8.15 (2022-07-20)
 ========================